:recycle: Separate escaping from `sb2html.ts`
